+++ play/src/front/Components/EmbedScreens/CamerasContainer.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import type { EmbedScreen } from "../../Stores/EmbedScreensStore";
+    import type { EmbedScreen } from "../../Stores/HighlightedEmbedScreenStore";
     import { streamableCollectionStore } from "../../Stores/StreamableCollectionStore";
     import MediaBox from "../Video/MediaBox.svelte";
     import { fly, fade } from "svelte/transition";
